feat: added BaseIntegrationTest
