+++ src/test/java/com/pawelnu/projectmanager/config/BaseIntegrationTest.java
@@ -1,0 +1,46 @@
+package com.pawelnu.projectmanager.config;
+
+import com.pawelnu.projectmanager.config.security.jwt.JwtUtils;
+import com.pawelnu.projectmanager.utils.Utils;
+import com.pawelnu.projectmanager.utils.Utils.Postgres;
+import com.pawelnu.projectmanager.utils.Utils.SpringDataSource;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.springframework.transaction.annotation.Transactional;
+import org.testcontainers.containers.PostgreSQLContainer;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@AutoConfigureMockMvc
+@ActiveProfiles("test")
+@Transactional
+public abstract class BaseIntegrationTest {
+
+  @Autowired protected JwtUtils jwtUtils;
+
+  static PostgreSQLContainer<?> postgres =
+      new PostgreSQLContainer<>(Postgres.POSTGRES_17)
+          .withDatabaseName(Postgres.DB_NAME)
+          .withUsername(Postgres.USER)
+          .withPassword(Postgres.PASSWORD);
+
+  @DynamicPropertySource
+  static void postgresProperties(DynamicPropertyRegistry registry) {
+    registry.add(SpringDataSource.URL, postgres::getJdbcUrl);
+    registry.add(SpringDataSource.USERNAME, postgres::getUsername);
+    registry.add(SpringDataSource.PASSWORD, postgres::getPassword);
+  }
+
+  static {
+    postgres.start();
+  }
+
+  @BeforeEach
+  void beforeEach() {
+    Utils.generateToken(jwtUtils);
+  }
+}
